requirements.txt
jinja2
playwright
python-dotenv
flask
flask-sqlalchemy
